[refactor] ControllerTransitionState: move to ui model as per ChatGPT's suggestion
